backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
chromadb>=0.5.0
python-dotenv>=1.0.1
google-genai>=0.2.0
pydantic>=2.7.4
